Revert "Change "Wireless ISP" to "wireless community-owned network.""
This reverts commit e3582c859dc2bbd35d07ced629d71e25ae7b0c9f.

This commit was originally made in response to the discussion in
nycmeshnet/docs#99. I'm not particularly married to either language, but
since the purpose of this fork is to publish what I feel is the best
possible version of the docs for the benefit of my community, I've
decided to revert to this language to meet that goal.

Concerns over the self-description of NYC Mesh is, bluntly, NYC Mesh's
responsibility and not mine. Especially after my experience in the
aforementioned discussion, I don't feel as though my time is best spent
engaging with the existing members of the NYC Mesh community for
substantial periods of time to discuss concerns that I see as entirely
theirs and not mine. I'd be very happy to see them base new versions of
their Docs on this fork, but the work that goes into this process to
make their version sound like theirs and not mine will be work they and
not I are responsible for doing.