Standardize on `slack.nycmesh.net` for the Slack home room.
This URL simply redirects to the "join" link on Slack for NYC Mesh. If
the user is already logged into the Slack group, the "join" link itself
redirects to the Slack app.

Note that the redirect on this subdomain does not yet understand
internal Slack URL endpoints such as `/messages` to load a single room,
so these URLs in the documentation remain as they are for now.